feat: convert from azure functions to vercel functions
