+++ .gitignore
@@ -70,3 +70,5 @@ __pycache__/
 __blobstorage__
 __queuestorage__
 __azurite_db*__.json
+
+*.vercel
