--- api/.funcignore
@@ -1,10 +1,0 @@
-*.js.map
-*.ts
-.git*
-.vscode
-local.settings.json
-test
-getting_started.md
-node_modules/@types/
-node_modules/azure-functions-core-tools/
-node_modules/typescript/
