+++ api/package.json
@@ -1,17 +1,5 @@
 {
-  "name": "",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/functions/*.js",
-  "scripts": {
-    "start": "func start",
-    "test": "echo \"No tests yet...\""
-  },
   "dependencies": {
-    "@azure/functions": "^4.0.0",
-    "axios": "^1.8.3"
-  },
-  "devDependencies": {
-    "azure-functions-core-tools": "^4.x"
+    "axios": "^1.8.4"
   }
 }
